Moved EventLoop to new file, started new mode impls
